+++ CHANGELOG.md
@@ -9,8 +9,9 @@ ADDED:
 - new class GenericTreeElement created
 - new class GenericTreeNodeCellRenderer created
 - new class TreeNodeGenericTreeElementWithContentPanel created
-- new class TreeNodeGenericTreeElementWithGenericContentPanel created
 - new method for get the selected user object from a given JTree object
+- new method for copy the selected DefaultMutableTreeNode object
+- new method factory method for create DefaultMutableTreeNode object 
 - new test dependency for icons org.freedesktop.tango:tango-icon-theme in patch version 0.8.90 
 - new test dependency for icons io.github.astrapi69:silk-icon-theme in major version 1
 
@@ -19,6 +20,7 @@ CHANGED:
 - update of gradle version to 7.3.3
 - update of gradle-plugin dependency com.github.ben-manes.versions.gradle.plugin to new version 0.41.0
 - update of dependency swing-base-components to new version 1.1
+- moved PreferencesPanel from swing-components to this repository
 
 Version 1
 -------------